requirements.txt
asgiref==3.7.2
Django==4.0.5
djangorestframework==3.13.1
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.3
drf-yasg==1.20.0
twilio==8.11.0
